Add dagshub_logger
