+++ pycaret/loggers/__init__.py
@@ -2,5 +2,12 @@ from .base_logger import BaseLogger
 from .dashboard_logger import DashboardLogger
 from .mlflow_logger import MlflowLogger
 from .wandb_logger import WandbLogger
+from .dagshub_logger import DagshubLogger
 
-__all__ = ["BaseLogger", "DashboardLogger", "MlflowLogger", "WandbLogger"]
+__all__ = [
+    "BaseLogger",
+    "DashboardLogger",
+    "MlflowLogger",
+    "WandbLogger",
+    "DagshubLogger",
+]
