+++ pycaret/loggers/mlflow_logger.py
@@ -1,5 +1,3 @@
-import datetime
-import os
 import secrets
 from copy import deepcopy
 
@@ -16,35 +14,13 @@ except ImportError:
 
 
 class MlflowLogger(BaseLogger):
-    def __init__(self, remote=None) -> None:
+    def __init__(self) -> None:
         if mlflow is None:
             raise ImportError(
                 "MlflowLogger requires mlflow. Install using `pip install mlflow`"
             )
         super().__init__()
         self.run = None
-        self.remote = remote
-
-        # Connect to repo
-        if self.remote:
-            try:
-                from dagshub.upload import Repo
-            except ImportError:
-                Repo = None
-
-            if Repo is None:
-                raise ImportError("mlflow remote server requires dagshub")
-            self.remote_model_root = "artifacts/models"
-            self.remote_rawdata_root = "artifacts/data/raw"
-            self.remote_procdata_root = "artifacts/data/process"
-            self.repo = Repo(
-                owner=os.getenv("REPO_OWNER"),
-                name=os.getenv("REPO_NAME"),
-                username=os.getenv("USER_NAME"),
-                password=os.getenv("PASSWORD"),
-                token=os.getenv("TOKEN"),
-                branch=os.getenv("BRANCH"),
-            )
 
     def init_experiment(self, exp_name_log, full_name=None):
         # get USI from nlp or tabular
@@ -57,8 +33,6 @@ class MlflowLogger(BaseLogger):
             except Exception:
                 pass
         full_name = full_name or f"{SETUP_TAG} {USI}"
-        if self.remote:
-            mlflow.set_tracking_uri(self.remote)
         mlflow.set_experiment(exp_name_log)
         self.run = mlflow.start_run(run_name=full_name, nested=True)
 
@@ -102,47 +76,7 @@ class MlflowLogger(BaseLogger):
         mlflow.set_tag("Run ID", RunID)
 
     def log_artifact(self, file, type="artifact"):
-        if self.remote:
-            if type == "model":
-                if not file.endswith("Transformation Pipeline.pkl"):
-                    remote_filename = os.path.join(
-                        self.remote_model_root, file
-                    )
-                    self.repo.upload(
-                        file=file,
-                        path=remote_filename,
-                        versioning="dvc",
-                        commit_message="update new trained model",
-                        force=True
-                    )
-            elif type in [
-                "train_data_remote",
-                "train_transform_data_remote",
-                "test_data_remote",
-                "test_transform_data_remote",
-            ]:
-                data_type = type.split("_")[0].lower()
-                is_transformed = "transform" in type
-                transformed = "transformed " if is_transformed else ""
-                remote_dir = (
-                    self.remote_procdata_root
-                    if is_transformed
-                    else self.remote_rawdata_root
-                )
-                remote_filename = os.path.join(
-                    remote_dir, file.split(os.sep)[-1]
-                )
-                self.repo.upload(
-                    file=file,
-                    path=remote_filename,
-                    versioning="dvc",
-                    commit_message=f"update {transformed}{data_type} data",
-                    force=True
-                )
-            else:
-                mlflow.log_artifact(file)
-        else:
-            mlflow.log_artifact(file)
+        mlflow.log_artifact(file)
 
     def log_plot(self, plot, title=None):
         self.log_artifact(plot)
